修复普通攻击的bug

+++ Packages/cn.etetet.wow/Scripts/Model/Share/Map/Spell/BT/TargetSelect/TargetSelectorSingle.cs
@@ -7,7 +7,7 @@
         [BTOutput(typeof(Unit))]
         public string Unit = BTEvnKey.Unit;
         
-        public float MaxDistance;
+        public int MaxDistance;
         
         public UnitType UnitType;
     }
